Day 9 solution

+++ src/main/kotlin/02-1-Program-Alarm.kt
@@ -5,7 +5,7 @@ import java.io.File
 
 fun main() {
     val input = File("inputs/02-1-input.txt").readText()
-    val program = input.split(",").map { it.toInt() }
+    val program = input.split(",").map { it.toLong() }
 
     val result = processProgram(program)
