+++ src/main/kotlin/05-2-Sunny-Asteroids.kt
@@ -5,7 +5,7 @@ import java.io.File
 
 fun main() {
     val input = File("inputs/05-1-input.txt").readText()
-    val program = input.split(",").map { it.toInt() }
+    val program = input.split(",").map { it.toLong() }
 
     val result = processProgram(program, intcodeInputs = listOf(5))
     println("Program outputs: ${result.outputs}")
