+++ src/main/kotlin/07-2-Amplification-Circuit.kt
@@ -5,7 +5,7 @@ import io.cloudtamer.adventofcode.lib.day7.optimizeProgram
 import java.io.File
 
 fun main() {
-    val inputProgram = File("inputs/07-1-input.txt").readText().split(",").map { it.toInt() }
+    val inputProgram = File("inputs/07-1-input.txt").readText().split(",").map { it.toLong() }
     val bestOutput = optimizeProgram(inputProgram, listOf(5, 6, 7, 8, 9), ::feedbackAmplify)
 
     println("Highest output: ${bestOutput.maxResult}")
