+++ src/main/kotlin/lib/day7/Part1.kt
@@ -2,12 +2,12 @@ package io.cloudtamer.adventofcode.lib.day7
 
 import io.cloudtamer.adventofcode.lib.day2.processProgram
 
-data class OptimizedInput(val inputSet: List<Int>, val maxResult: Int)
+data class OptimizedInput(val inputSet: List<Long>, val maxResult: Long)
 
 fun optimizeProgram(
-    program: List<Int>,
-    possibleInputs: List<Int>,
-    amplifyAlgorithm: (List<Int>, List<Int>) -> Int
+    program: List<Long>,
+    possibleInputs: List<Long>,
+    amplifyAlgorithm: (List<Long>, List<Long>) -> Long
 ): OptimizedInput {
     val inputPermutations = generatePermutations(possibleInputs)
 
@@ -17,19 +17,19 @@ fun optimizeProgram(
     return OptimizedInput(maxInput, maxResult)
 }
 
-fun amplify(program: List<Int>, inputSet: List<Int>): Int {
-    var lastOutput = 0
+fun amplify(program: List<Long>, inputSet: List<Long>): Long {
+    var lastOutput = 0L
     for (input in inputSet) {
         lastOutput = processProgram(program, listOf(input, lastOutput)).firstOutput ?: throw IllegalStateException("Amplifier didn't produce output!")
     }
     return lastOutput
 }
 
-fun generatePermutations(toSelect: List<Int>, selected: List<Int> = emptyList()): List<List<Int>> {
+fun generatePermutations(toSelect: List<Long>, selected: List<Long> = emptyList()): List<List<Long>> {
     return if (toSelect.isEmpty()) {
         listOf(selected)
     } else {
-        val permutations = mutableListOf<List<Int>>()
+        val permutations = mutableListOf<List<Long>>()
         for (option in toSelect) {
             permutations += generatePermutations(toSelect - option, selected + option)
         }
